pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "treebranchmarks"
version = "0.1.0"
description = "Benchmarking framework for comparing SHAP algorithms on decision tree ensembles"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "xgboost>=2.0",
    "lightgbm>=4.0",
    "shap>=0.51",
    "joblib>=1.3",
    "plotly>=5.18",
    "requests>=2.31",
    "woodelf_explainer",
    "pyarrow",
    "fastparquet"
]

[project.optional-dependencies]
kaggle = ["kaggle>=1.5"]
openml = ["openml>=0.14"]
dev = ["pytest>=7.4", "pytest-cov"]

[tool.setuptools.packages.find]
where = ["."]
include = ["treebranchmarks*"]
